pyproject.toml
[tool.poetry]
name = "political-discourse-analyzer"
version = "0.1.0"
description = "Herramienta de análisis de discursos políticos mediante IA"
authors = ["Angélica Laza <angi.laza@hotmail.es>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.24.0"
langchain = "^0.1.0"
openai = "^1.0.0"
pydantic = "^2.6.0"
python-dotenv = "^1.0.0"
PyPDF2 = "^3.0.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^24.1.0"
mypy = "^1.8.0"
ruff = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"